po and material issue chnages

+++ app/Http/Controllers/MaterialIssueController.php
@@ -1,21 +1,20 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use App\Models\MaterialIssue;
 use App\Models\MaterialIssueItem;
 use App\Models\ProjectMaster;
 use App\Models\SiteMaster;
 use App\Models\ItemMaster;
 use App\Models\EmployeeMaster;
+use Exception;
+
 use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Eloquent\Model;
 use App\Http\Requests\MaterialIssueRequest;
-use Exception;
 
 
 use Illuminate\Http\Request;
-
 require_once(app_path('constants.php'));
 
 class MaterialIssueController extends Controller
@@ -25,32 +24,55 @@ class MaterialIssueController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    // autocomplete data from site name
+
+    //don't touch that
+    public function  getitemnamedata(){
+      try{
+        $itemname = $_GET['itemname'];
+        info($itemname);
+        $data = ItemMaster::where('item_name','LIKE',$itemname.'%')->get();
+        info($data);
+
+        return $data;
+    }
+     catch (Exception $e) {
+        info($e);
+        return response()->json('Error occured in the loading page', 400);
+    }
+}
+    public function  getsitelocationdata(){
+        $site_name = $_GET['site_name'];
+
+        $data = SiteMaster::where('site_location','LIKE',$site_name.'%')->get();
+        //info($data);
+
+        return $data;
+    }
+
     public function index()
     {
-        try {
+        try{
             $type = MATERIALTYPE;
 
-            $material_issues = DB::table('material_issue_return')
-                ->join('employee_masters', 'material_issue_return.receiving_employee', '=', 'employee_masters.id')
-                ->join('project_masters', 'material_issue_return.project_no', '=', 'project_masters.project_no')
-                ->select(
-                    'employee_masters.*',
-                    'project_masters.*',
-                    'material_issue_return.*',
-                    DB::raw('DATE(material_issue_return.issue_date) as issue_date')
-                )
-                ->get();
-            return view('materialissue.index')->with([
-                'material_issues' => $material_issues,
-                'type' => $type,
+        $material_issues = DB::table('material_issue_return')
+             ->join('employee_masters', 'material_issue_return.receiving_employee', '=', 'employee_masters.id')
+            ->join('project_masters', 'material_issue_return.project_no', '=', 'project_masters.project_no')
+            ->select( 'employee_masters.*','project_masters.*', 'material_issue_return.*',
+            DB::raw('DATE(material_issue_return.issue_date) as issue_date'))
+            ->get();
+        return view('materialissue.index')->with([
+            'material_issues' => $material_issues,
+            'type' => $type,
 
-            ]);
+        ]);
 
-        } catch (Exception $e) {
-            info($e);
-            return response()->json('Error occured in the loading page', 400);
-        }
     }
+    catch (Exception $e) {
+        info($e);
+        return response()->json('Error occured in the loading page', 400);
+    }
+}
 
     /**
      * Show the form for creating a new resource.
@@ -72,18 +94,18 @@ class MaterialIssueController extends Controller
     {
 
         try {
-
+            info($request);
 
             MaterialIssue::create($request->only(MaterialIssue::REQUEST_INPUTS));
-            $mir_no = MaterialIssue::max('mir_no');
-            for ($i = 0; $i < count($request['item_no']); $i++) {
+            $mir_no= MaterialIssue::max('mir_no');
+
+            for ($i = 0;$i <  count($request['item_no']); $i++) {
 
                 MaterialIssueItem::create([
 
-                    //'material_issue_id' => $request['material_issue_id'][$i],
                     'mir_no' => $mir_no,
                     'item_no' => $request['item_no'][$i],
-                    'item' => $request['item'][$i],
+                    // 'item' => $request['item'][$i],
                     'store_room' => $request['store_room'][$i],
                     'item_quantity' => $request['item_quantity'][$i],
                 ]);
@@ -105,36 +127,33 @@ class MaterialIssueController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($mir_no)
-    {
-        try {
-            $material_issues = DB::table('material_issue_return')
-                ->join('employee_masters', 'material_issue_return.receiving_employee', '=', 'employee_masters.id')
-                ->join('project_masters', 'material_issue_return.project_no', '=', 'project_masters.project_no')
+public function show($mir_no)
+{
+    try {
+        $material_issues = DB::table('material_issue_return')
+            ->join('employee_masters', 'material_issue_return.receiving_employee', '=', 'employee_masters.id')
+           ->join('project_masters', 'material_issue_return.project_no', '=', 'project_masters.project_no')
 
-                ->select(
-                    'employee_masters.*',
-                    'project_masters.*',
-                    'material_issue_return.*',
-                    DB::raw('DATE(material_issue_return.issue_date) as issue_date')
-                )
-                ->where('material_issue_return.mir_no', $mir_no)
-                ->get();
-            $material_issues_item = DB::table('material_issue_return_item')
-                ->join('item_masters', 'material_issue_return_item.item_no', '=', 'item_masters.id')
-                ->join('material_issue_return', 'material_issue_return_item.mir_no', '=', 'material_issue_return.mir_no')
-                ->select('material_issue_return_item.*', 'item_masters.*', 'material_issue_return.*')
-                ->where('material_issue_return_item.mir_no', $mir_no)
-                ->get();
-            return response()->json([
-                'material_issues' => $material_issues,
-                'material_issues_item' => $material_issues_item,
-            ]);
-        } catch (Exception $e) {
-            info($e);
-            return response()->json('Error occurred in the show', 400);
-        }
+           ->select( 'employee_masters.*','project_masters.*', 'material_issue_return.*',
+           DB::raw('DATE(material_issue_return.issue_date) as issue_date'))
+           ->where('material_issue_return.mir_no',$mir_no)
+           ->get();
+           $material_issues_item = DB::table('material_issue_return_item')
+           ->join('item_masters', 'material_issue_return_item.item_no', '=', 'item_masters.id')
+           ->join('material_issue_return', 'material_issue_return_item.mir_no', '=', 'material_issue_return.mir_no')
+            ->select('material_issue_return_item.*', 'item_masters.*','material_issue_return.*')
+            ->where('material_issue_return_item.mir_no', $mir_no)
+            ->get();
+
+        return response()->json([
+            'material_issues' => $material_issues,
+            'material_issues_item'=>$material_issues_item,
+        ]);
+    } catch (Exception $e) {
+        info($e);
+        return response()->json('Error occurred in the show', 400);
     }
+}
 
 
     /**
@@ -145,7 +164,7 @@ class MaterialIssueController extends Controller
      */
     public function edit($id)
     {
-        //
+       //
     }
 
 
@@ -157,8 +176,9 @@ class MaterialIssueController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-    public function update(Request $request, $mir_no)
+       public function update(Request $request, $mir_no)
     {
+
         try {
             DB::beginTransaction();
 
@@ -167,18 +187,17 @@ class MaterialIssueController extends Controller
 
             // update the Material Issue item data
             $itemCount = count($request['item_no']);
-            $mir_delete = MaterialIssueItem::where('mir_no', $mir_no)->delete();
+            $mir_delete=MaterialIssueItem::where('mir_no',$mir_no)->delete();
             for ($i = 0; $i < $itemCount; $i++) {
                 MaterialIssueItem::create([
-                    'mir_no' => $mir_no,
-                    'item' => $request['item'][$i],
+                    'mir_no'=>$mir_no,
+                    // 'item' => $request['item'][$i],
                     'item_no' => $request['item_no'][$i],
                     'store_room' => $request['store_room'][$i],
                     'item_quantity' => $request['item_quantity'][$i],
                 ]);
-                // info($material_issues_item);
-            }
 
+            }
             DB::commit();
             return response()->json('Material Updated Successfully', 200);
         } catch (Exception $e) {
@@ -197,20 +216,40 @@ class MaterialIssueController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-    public function destroy($mir_no)
-    {
+     public function destroy($mir_no){
         try {
-            $material_issues_item = MaterialIssueItem::where('mir_item_no', $mir_no)->first();
-            if ($material_issues_item != null) {
-                $material_issues_item->delete();
+            // Disable foreign key check
+            DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+
+            $material_issues = MaterialIssue::where('mir_no', $mir_no)->first();
+
+            if ($material_issues != null) {
+                $material_issues_items = MaterialIssueItem::where('mir_item_no', $mir_no)->get();
+
+                foreach ($material_issues_items as $material_issues_item) {
+                    $material_issues_item->delete();
+                }
+
+                $material_issues->delete();
+
+                // Enable foreign key check
+                DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+
+                return response()->json('Material Deleted Successfully', 200);
+            } else {
+                // Enable foreign key check
+                DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+
+                return response()->json('Material not found', 404);
             }
-            $material_issues = MaterialIssue::where('mir_no', $mir_no);
-            $material_issues->delete();
-            return response()->json('Material Deleted Successfully', 200);
         } catch (Exception $e) {
+            // Enable foreign key check
+            DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+
             info($e);
-            return response()->json('Error occured in the delete', 400);
+            return response()->json('Error occurred in the delete', 400);
         }
     }
+
 
 }
